Decouple evaluation and execution in the preemption framework
